+++ docs/source/conf.py
@@ -1,4 +1,7 @@
 from h3ronpy import __version__ as hp_version
+import os
+from pathlib import Path
+
 
 # Configuration file for the Sphinx documentation builder.
 
@@ -41,6 +44,4 @@ epub_show_urls = "footnote"
 
 autodoc_typehints = "both"
 
-import os
-from pathlib import Path
 os.environ["PROJECT_ROOT"] = str(Path(__file__).parent.parent.parent)